Fix a deamon crash

+++ compile/compile.go
@@ -73,14 +73,18 @@ func home(w http.ResponseWriter, r *http.Request) {
 			if ( len(device) > 1 ) {
 				fmt.Printf("Device: %d\n", device)
 				if ( device == "bmc" ) {
-					unix.Kill(OpenBMCCommand.Process.Pid, unix.SIGINT)
-	                                _ = <- OpenBMCBuildChannel
-	                                OpenBMCCommand = nil
+					if ( OpenBMCCommand != nil ) {
+						unix.Kill(OpenBMCCommand.Process.Pid, unix.SIGINT)
+		                                _ = <- OpenBMCBuildChannel
+		                                OpenBMCCommand = nil
+					}
 				} else {
 					if ( device == "rom" ) {
-						unix.Kill(LinuxBOOTCommand.Process.Pid, unix.SIGINT)
-		                                _ = <- LinuxBOOTBuildChannel
-		                                LinuxBOOTCommand = nil
+						if ( LinuxBOOTCommand != nil ) {
+							unix.Kill(LinuxBOOTCommand.Process.Pid, unix.SIGINT)
+			                                _ = <- LinuxBOOTBuildChannel
+			                                LinuxBOOTCommand = nil
+						}
 					}
 				}
 			}
